requirements.txt
Pillow
pyyaml>=5.1
sentencepiece
tensorboard
tensorboardX
textgrid
pytest
flake8==3.8.2
flake8-bugbear
flake8-comprehensions
flake8-executable
flake8-pyi==20.5.0
mccabe
pycodestyle==2.6.0
pyflakes==2.2.0
clang-format==17.0.6
cpplint==1.6.1
torch>=2.1.2
torchaudio>=2.1.2
tqdm
deepspeed>=0.14.0
librosa
openai-whisper==20231117
pre-commit==3.5.0
langid
